requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-dotenv
pydantic_settings
python-jose[cryptography]
passlib
passlib[bcrypt]
firebase-admin
requests
aioredis
websockets